Last updates

+++ services/booking-service/src/main/java/com/culturalspace/bookingservice/event/BookingConfirmedEvent.java
@@ -7,12 +7,12 @@ public class BookingConfirmedEvent {
     private Long bookingId;
     private Long userId;
     private int numberOfSeats;
-    private BigDecimal totalPrice;
+    private double totalPrice;
 
     // Default constructor is required for JSON deserialization by Kafka/Spring
     public BookingConfirmedEvent() {}
 
-    public BookingConfirmedEvent(Long bookingId, Long userId, int numberOfSeats, BigDecimal totalPrice) {
+    public BookingConfirmedEvent(Long bookingId, Long userId, int numberOfSeats, double totalPrice) {
         this.bookingId = bookingId;
         this.userId = userId;
         this.numberOfSeats = numberOfSeats;
@@ -23,14 +23,14 @@ public class BookingConfirmedEvent {
     public Long getBookingId() { return bookingId; }
     public Long getUserId() { return userId; }
     public int getNumberOfSeats() { return numberOfSeats; }
-    public BigDecimal getTotalPrice() { return totalPrice; }
+    public double getTotalPrice() { return totalPrice; }
 
     // --- Setters ---
     // (Setters are often needed for deserialization, though not always used if data is immutable)
     public void setBookingId(Long bookingId) { this.bookingId = bookingId; }
     public void setUserId(Long userId) { this.userId = userId; }
     public void setNumberOfSeats(int numberOfSeats) { this.numberOfSeats = numberOfSeats; }
-    public void setTotalPrice(BigDecimal totalPrice) { this.totalPrice = totalPrice; }
+    public void setTotalPrice(double totalPrice) { this.totalPrice = totalPrice; }
 
     @Override
     public String toString() {
